# text-mine-backend/src/main/resources/generated/api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "text-mine",
    "version" : "v1",
    "title" : "text-mine-api"
  },
  "host" : "localhost:8080",
  "basePath" : "/",
...
  "schemes" : [ "http" ],
  "paths" : {
    "/articles" : {
      "get" : {
        "tags" : [ "articles" ],
        "summary" : "Get filtered articles",
        "description" : "",
        "operationId" : "getFiltered",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "entryWord",
          "in" : "query",
          "description" : "Entry word",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "formVariant",
          "in" : "query",
          "description" : "Form variant",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "inflection",
          "in" : "query",
          "description" : "Inflection",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "pos",
          "in" : "query",
          "description" : "Parts of speech",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        }, {
          "name" : "documentId",
          "in" : "query",
          "description" : "Document ids",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        }, {
          "name" : "corpusId",
          "in" : "query",
          "description" : "Corpus id",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "matchingStrategy",
          "in" : "query",
          "description" : "Matching strategy",
          "required" : false,
          "type" : "string",
          "enum" : [ "STARTS_WITH", "ENDS_WITH", "CONTAINS", "EXACT_MATCH" ]
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "Offset",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "Limit",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ArticleDTO"
              }
            }
          }
        }
      }
    },
    "/articles/by-document/{documentId}" : {
      "get" : {
        "tags" : [ "articles" ],
        "summary" : "Get articles of a document",
        "description" : "",
        "operationId" : "getByDocument",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "documentId",
          "in" : "path",
          "description" : "Document Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "section",
          "in" : "query",
          "description" : "Section number",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page number",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ArticleDTO"
              }
            }
          }
        }
      }
    },
    "/articles/group/by-document" : {
      "get" : {
        "tags" : [ "articles" ],
        "summary" : "Get articles grouped by document id",
        "description" : "",
        "operationId" : "groupByDocument",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "entryWord",
          "in" : "query",
          "description" : "Entry word",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "formVariant",
          "in" : "query",
          "description" : "Form variant",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "inflection",
          "in" : "query",
          "description" : "Inflection",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "pos",
          "in" : "query",
          "description" : "Parts of speech",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        }, {
          "name" : "corpusId",
          "in" : "query",
          "description" : "Corpus Id",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "matchingStrategy",
          "in" : "query",
          "description" : "Matching strategy",
          "required" : false,
          "type" : "string",
          "enum" : [ "STARTS_WITH", "ENDS_WITH", "CONTAINS", "EXACT_MATCH" ]
        }, {
          "name" : "posCount",
          "in" : "query",
          "description" : "Part of speech count",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "$ref" : "#/definitions/ArticleDTOList"
              }
            }
          }
        }
      }
    },
    "/articles/{id}" : {
      "get" : {
        "tags" : [ "articles" ],
        "summary" : "Get article by Id",
        "description" : "",
        "operationId" : "getOne",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Article Id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ArticleDTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "articles" ],
        "summary" : "Update article",
        "description" : "",
        "operationId" : "putArticle",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Article",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Article"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Article id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ArticleDTO"
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "articles" ],
        "summary" : "Delete article",
        "description" : "",
        "operationId" : "deleteArticle",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Article Id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/documents" : {
      "get" : {
        "tags" : [ "documents" ],
        "summary" : "Get filtered documents",
        "description" : "",
        "operationId" : "getFiltered",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "entryText",
          "in" : "query",
          "description" : "Entry text",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "pos",
          "in" : "query",
          "description" : "Parts of speech",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        }, {
          "name" : "entryWord",
          "in" : "query",
          "description" : "Entry words",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        }, {
          "name" : "documentId",
          "in" : "query",
          "description" : "Document ids",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "Offset",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "Limit",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Document"
              }
            }
          }
        }
      }
    },
    "/documents/attachment/{attachmentId}" : {
      "get" : {
        "tags" : [ "documents" ],
        "summary" : "Get attachment by id",
        "description" : "",
        "operationId" : "getFile",
        "produces" : [ "application/octet-stream" ],
        "parameters" : [ {
          "name" : "attachmentId",
          "in" : "path",
          "description" : "Attachment id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string",
                "format" : "byte"
              }
            }
          }
        }
      }
    },
...
    "/notes/by-document/{id}" : {
      "get" : {
        "tags" : [ "notes" ],
        "summary" : "Get note by document id",
        "description" : "",
        "operationId" : "getByDocument",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Document id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Note"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Article" : {
      "type" : "object",
      "required" : [ "derivative", "document", "entryWord", "iri", "properNoun" ],
      "properties" : {
        "document" : {
          "$ref" : "#/definitions/Document"
        },
        "entryWord" : {
          "type" : "string"
        },
        "properNoun" : {
          "type" : "boolean"
        },
        "derivative" : {
          "type" : "boolean"
        },
        "editorNote" : {
          "type" : "string"
        },
        "meaning" : {
          "type" : "string"
        },
        "partOfSpeech" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "enum" : [ "PRONOMINAL_ADVERB", "PRONOUN", "NOUN", "INFINITIVE", "ADVERBIAL_PARTICIPLE", "ARTICLE", "ADVERB", "NUMERAL", "VERB", "VERB_PARTICIPLE", "PREVERB", "INTERJECTION", "CONJUNCTIVE", "ADJECTIVE", "PARTICIPLE", "MODIFIER", "SENTENTIAL", "POSTPOSITION", "POSTPOSITIONAL_ADJECTIVE", "PARTICLE", "AUXILIARY" ]
          }
        },
        "externalReferences" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "internalReferences" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "formVariants" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/FormVariant"
          }
        },
        "score" : {
          "type" : "number",
          "format" : "double"
        },
        "iri" : {
          "type" : "string"
        },
        "id" : {
          "$ref" : "#/definitions/ObjectId"
        },
        "version" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastModifiedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "createdBy" : {
          "type" : "string"
        },
        "lastModifiedBy" : {
          "type" : "string"
        },
        "hash" : {
          "type" : "string"
        }
      }
    },
    "ArticleDTO" : {
      "type" : "object",
      "properties" : {
        "documentId" : {
          "type" : "string"
        },
        "entryWord" : {
          "type" : "string"
        },
        "properNoun" : {
          "type" : "boolean"
        },
        "derivative" : {
          "type" : "boolean"
        },
        "editorNote" : {
          "type" : "string"
        },
        "meaning" : {
          "type" : "string"
        },
        "partOfSpeech" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "enum" : [ "PRONOMINAL_ADVERB", "PRONOUN", "NOUN", "INFINITIVE", "ADVERBIAL_PARTICIPLE", "ARTICLE", "ADVERB", "NUMERAL", "VERB", "VERB_PARTICIPLE", "PREVERB", "INTERJECTION", "CONJUNCTIVE", "ADJECTIVE", "PARTICIPLE", "MODIFIER", "SENTENTIAL", "POSTPOSITION", "POSTPOSITIONAL_ADJECTIVE", "PARTICLE", "AUXILIARY" ]
          }
        },
        "externalReferences" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "internalReferences" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "formVariants" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/FormVariant"
          }
        },
        "id" : {
          "$ref" : "#/definitions/ObjectId"
        },
        "version" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastModifiedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "createdBy" : {
          "type" : "string"
        },
        "lastModifiedBy" : {
          "type" : "string"
        }
      }
    },
    "ArticleDTOList" : {
      "type" : "object",
      "properties" : {
        "size" : {
          "type" : "integer",
          "format" : "int64"
        },
        "rangeFrom" : {
          "type" : "integer",
          "format" : "int64"
        },
        "rangeSize" : {
          "type" : "integer",
          "format" : "int64"
        },
        "entries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ArticleDTO"
          }
        }
      }
    },
    "Corpus" : {
      "type" : "object",
      "required" : [ "description", "iri", "title" ],
      "properties" : {
        "iri" : {
          "type" : "string"
        },
        "id" : {
          "$ref" : "#/definitions/ObjectId"
        },
        "version" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastModifiedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "createdBy" : {
          "type" : "string"
        },
        "lastModifiedBy" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "hash" : {
          "type" : "string"
        }
      }
    },
    "Document" : {
      "type" : "object",
      "required" : [ "attachmentIds", "author", "content", "corpus", "iri", "pages", "sections", "title" ],
      "properties" : {
        "corpus" : {
          "$ref" : "#/definitions/Corpus"
        },
        "title" : {
          "type" : "string"
        },
        "content" : {
          "type" : "string"
        },
        "normalized" : {
          "type" : "string"
        },
        "author" : {
          "type" : "string"
        },
        "sections" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Section"
          }
        },
        "pages" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Section"
          }
        },
        "attachmentIds" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "score" : {
          "type" : "number",
          "format" : "double"
        },
        "iri" : {
          "type" : "string"
        },
        "id" : {
          "$ref" : "#/definitions/ObjectId"
        },
        "version" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastModifiedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "createdBy" : {
          "type" : "string"
        },
        "lastModifiedBy" : {
          "type" : "string"
        },
        "hash" : {
          "type" : "string"
        }
      }
    },
...
    "FormVariant" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "inflections" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Inflection"
          }
        }
      }
    },
    "Inflection" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "occurrences" : {
          "type" : "integer",
          "format" : "int32"
        },
        "examples" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/EntryExample"
          }
        }
      }
    },
...
    "Note" : {
      "type" : "object",
      "required" : [ "content", "documentId", "iri", "lineRefs", "quote", "section", "subType", "type" ],
      "properties" : {
        "documentId" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "subType" : {
          "type" : "string"
        },
        "quote" : {
          "type" : "string"
        },
        "lineRefs" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "section" : {
          "type" : "integer",
          "format" : "int32"
        },
        "content" : {
          "type" : "string"
        },
        "score" : {
          "type" : "number",
          "format" : "double"
        },
        "iri" : {
          "type" : "string"
        },
        "id" : {
          "$ref" : "#/definitions/ObjectId"
        },
        "version" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastModifiedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "createdBy" : {
          "type" : "string"
        },
        "lastModifiedBy" : {
          "type" : "string"
        },
        "hash" : {
          "type" : "string"
        }
      }
    },
...
    "ObjectId" : {
      "type" : "object",
      "properties" : {
        "timestamp" : {
          "type" : "integer",
          "format" : "int32"
        },
        "machineIdentifier" : {
          "type" : "integer",
          "format" : "int32"
        },
        "processIdentifier" : {
          "type" : "integer",
          "format" : "int32"
        },
        "counter" : {
          "type" : "integer",
          "format" : "int32"
        },
        "time" : {
          "type" : "integer",
          "format" : "int64"
        },
        "date" : {
          "type" : "string",
          "format" : "date-time"
        },
        "timeSecond" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "Section" : {
      "type" : "object",
      "required" : [ "iri" ],
      "properties" : {
        "serial" : {
          "type" : "integer",
          "format" : "int32"
        },
        "content" : {
          "type" : "string"
        },
        "normalized" : {
          "type" : "string"
        },
        "lines" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Line"
          }
        },
        "iri" : {
          "type" : "string"
        },
        "hash" : {
          "type" : "string"
        }
      }
    },
...
  }
}